% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_abs_datierungen.R
\name{absolute_datierungen}
\alias{absolute_datierungen}
\title{Adds absolute datings}
\usage{
absolute_datierungen(absolute_data_list, liste_stellen, matrix_gleich,
  matrix_ueber_unter)
}
\arguments{
\item{absolute_data_list}{PLATZHALTERPLATZHALTER}

\item{liste_stellen}{PLATZHALTERPLATZHALTER}

\item{matrix_gleich}{PLATZHALTERPLATZHALTER}

\item{matrix_ueber_unter}{PLATZHALTERPLATZHALTER}
}
\value{
PLATZHALTER PLATZHALTERPLATZHALTERPLATZHALTER
}
\description{
Adds absolute datings
}
\examples{
absolute_chronology <- absolute_datierungen(absolute_data_list, liste_stellen, matrix_gleich, matrix_ueber_unter)
}
