% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_stratigrafie_programm.R
\name{RStratigraphy}
\alias{RStratigraphy}
\title{Stratifying archaeological data}
\usage{
RStratigraphy(strat_list, absolute_data_list, fehler_loeschen)
}
\arguments{
\item{strat_list}{A table of three columns and n rows, containing all known relations between the features. The first row has to name the columns. The first and the third columns contains the names of the features, where as the second column containes the relation in form of "above", "under" or "equal".}

\item{absolute_data_list}{A table of two columns and n rows, containing all known absolute datings of dated features. The first row has to name the columns. The first column contains the name of the feature, the second one the datings as numeric number.}

\item{fehler_loeschen}{A function, that automaticly erases logical mistakes in your stratigraphy if set on TRUE. WARNING! This function does not establish scientifically correct data!}
}
\value{
A list of tables: under-above-relation of the features = tab_under_above, equal features = tab_equal, absolute datings of features = absolute_chronology.
}
\description{
\code{RStratigraphy} returns different tables in which logical relations between a stratified data are completed, as well as their absolute datings. In addition, logical mistakes in the dataset are revealed.
}
\examples{
RStratigraphy <- function(strat_list, absolute_data_list, fehler_loeschen)
}
